Add `name` computed property to `ServiceContextKey` extension (#67)
This exposes the `AnyServiceContextKey.name` property directly on
`ServiceContextKey`.

Without this, downstream projects could not access the key name easily
because the `AnyServiceContextKey` initializer is internal. The only
workarounds were to copy the implementation of the `name` property, or
to add the key to a `ServiceContext` and then use `forEach` to get the
`AnyServiceContextKey` and access its `name`.